refactor : LottoResults 클래스 생성 및 전체적인 구조 수정

+++ src/main/java/domain/Rank.java
@@ -41,7 +41,7 @@ public enum Rank {
         return rank != OTHER;
     }
 
-    public int getAmount() {
+    public long getAmount() {
         return amount;
     }
